fix: extract system settings registration logic into a standalone module
