+++ module/hooks/i18n.mjs
@@ -1,6 +1,6 @@
-import registerSystemSettings from "../helpers/settings.mjs";
 import * as data from "../data/_module.mjs";
 import { performPreLocalization } from "../i18n/localization.mjs";
+import registerSystemSettings from "../system/settings.mjs";
 
 /**
  *
